+++ mail/mail-common/src/mail_scroller/mail_scroller_source/remote_source/search_scroller_source.rs
@@ -395,6 +395,13 @@ impl MailScrollerSource for SearchScrollerSource {
         }
     }
 
+    async fn sync_new(
+        &mut self,
+        _ctx: &MailUserContext,
+    ) -> Result<MailPaginatorJoinHandle, MailContextError> {
+        Ok(None)
+    }
+
     async fn clear_cursor(
         &mut self,
         ctx: &MailUserContext,