+++ mail/mail-common/src/models/message.rs
@@ -1314,6 +1314,7 @@ impl Message {
         WHERE
           expiration_time < STRFTIME('%s', 'NOW')
           AND expiration_time != 0
+          AND deleted = 0
         ",
             vec![],
             tether,